Added corpus
# Conflicts:
#	kzlib/jvm/src/test/kotlin/com/soywiz/kzlib/KZlibRefJZlib.kt

+++ kzlib/jvm/src/test/java/com/jcraft/jzlib/Checksum.java
@@ -35,9 +35,13 @@ EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 package com.jcraft.jzlib;
 
 interface Checksum {
-  void update(byte[] buf, int index, int len);
-  void reset();
-  void reset(long init);
-  long getValue();
-  Checksum copy();
+	void update(byte[] buf, int index, int len);
+
+	void reset();
+
+	void reset(int init);
+
+	int getValue();
+
+	Checksum copy();
 }
